Refactor project structure and update dependencies
