+++ src/lib/design-tokens.ts
@@ -1,0 +1,9 @@
+// ────────────────────────────────────────────────────────────────────────────────
+// DESIGN TOKENS - COMPATIBILITY EXPORT
+// ────────────────────────────────────────────────────────────────────────────────
+
+// Re-export everything from the actual design tokens location
+export * from '../design-system/tokens/design-tokens';
+
+// Explicit re-exports for compatibility
+export { typographyTokens, colors, designTokens, spacing, borderRadius, shadows } from '../design-system/tokens/design-tokens';
